Use mergo, move to other config helpers
